Add tests for starter validation and component registration
